# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP API for the research monitor (arXiv ingestion + paper catalogue).",
        "title": "Research Monitor API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/sources": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sources"
                ],
                "summary": "List all sources",
                "responses": {
                    "200": {
                        "description": "Sources",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller.SourceListEnvelope"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API token",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sources"
                ],
                "summary": "Create a source",
                "parameters": [
                    {
                        "description": "Source to create",
                        "name": "source",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_domain_source.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created source",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller.SourceEnvelope"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON body or validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API token",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "Source URL already exists",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/sources/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sources"
                ],
                "summary": "Get a source by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Source",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller.SourceEnvelope"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API token",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Source not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Sources"
                ],
                "summary": "Delete a source",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "401": {
                        "description": "Missing or invalid API token",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Source not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sources"
                ],
                "summary": "Update a source",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "source",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_domain_source.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated source",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller.SourceEnvelope"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON body or validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API token",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Source not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_yoavweber_research-monitor_backend_internal_domain_source.CreateRequest": {
            "type": "object",
            "required": [
                "kind",
                "name",
                "url"
            ],
            "properties": {
                "kind": {
                    "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_domain_source.Kind"
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "github_com_yoavweber_research-monitor_backend_internal_domain_source.Kind": {
            "type": "string",
            "enum": [
                "rss",
                "api"
            ],
            "x-enum-varnames": [
                "KindRSS",
                "KindAPI"
            ]
        },
        "github_com_yoavweber_research-monitor_backend_internal_domain_source.Response": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "kind": {
                    "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_domain_source.Kind"
                },
                "last_fetched_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "github_com_yoavweber_research-monitor_backend_internal_domain_source.UpdateRequest": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_yoavweber_research-monitor_backend_internal_http_common.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.Error"
                }
            }
        },
...
        "internal_http_controller.SourceEnvelope": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_domain_source.Response"
                }
            }
        },
        "internal_http_controller.SourceListEnvelope": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_domain_source.Response"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT issued by POST /auth/login. Supply as \"Bearer \u0026lt;token\u0026gt;\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}